* origin/release/v1.2.0: (24 commits)
  Prepare release v1.2.0-rc1.
  Fix version string in documentation configuration.
  Added Changelog file.
  Bump version (v1.2.0)
  Example : Code refactor of example 07.
  Revert v1.0.0 : Remove number of entries in Attribute Reporting reports.
  Documentation fixes.
  Added data size checks to Serializable implementations.
  Print stack trace on assert when running unit tests.
  Minor code format and documentation fixes.
  Updating CppUTest test library.
  Added checks on SerializableHelper pack and unpack.
  Added helper method to Common::ByteArray.
  Change size_t usage to uint16_t.
  Bug fix in AbstractInterface::check_payload_size.
  Revert v1.0.0 : Fixes in Attribute Reporting service.
  Minor fix to comments.
  Revert v1.0.0 : Fixed Device Information attributes.
  Revert v1.0.0 : Merged UID's RFPI and IPUI.
  Revert v1.0.0 : Remove address from de-register response.
  ...

+++ apps/base.h
@@ -5,7 +5,7 @@
  * This file contains the definition of the Base class that represents the
  * HAN-FUN Concentrator on the application.
  *
- * @version    1.1.1
+ * @version    1.2.0
  *
  * @copyright  Copyright &copy; &nbsp; 2014 ULE Alliance
  *
@@ -205,7 +205,7 @@ struct Base:public HF::Devices::Concentrator::Abstract <Unit0>
    virtual ~Base()
    {}
 
-   void receive (HF::Protocol::Packet &packet, HF::Common::ByteArray &payload, size_t offset);
+   void receive (HF::Protocol::Packet &packet, HF::Common::ByteArray &payload, uint16_t offset);
 
    using HF::Devices::Concentrator::Abstract <Unit0>::link;
